Add Whatsapp support
This will add Whatsapp support.

Given Whatsapp doesn't have a desktop web sharing URL like other
services, we have to do some out-of-the-box implementation here.

Whatsapp sharing URL starts with `whatsapp://send`, not `https://`,
which works on mobile but not on desktop (unless you have the Whatsapp Desktop
app). That means it will break for the majority of people on desktop.

To minimize risks, we:

1. don't add `whatsapp` by default and leave this decision up
to the developer. If they want to add it, they just need to add `whatsapp` to
the `allow_sites` config initializer.
2. we have a new CSS rule that will hide it on desktop browsers.
Developers are free to override it if they want to show it everywhere.

Extra: added a pt-BR locale file.